+++ R/parse_front_matter.R
@@ -106,15 +106,45 @@ parse_front_matter <- function(text, parse_yaml = NULL, parse_toml = NULL) {
   }
 
   parsed_data <- switch(
-    result$fence_type,
+    result$format,
     yaml = parse_yaml(result$content),
     toml = parse_toml(result$content),
     NULL
   )
 
-  list(
+  ret <- list(
     data = parsed_data,
     body = result$body
+  )
+  attr(ret, "format") <- result$format
+  attr(ret, "fence_type") <- result$fence_type
+
+  structure(ret, class = "front_matter")
+}
+
+#' @export
+print.front_matter <- function(x, ...) {
+  cat(sprintf(
+    "<front_matter format=\"%s\", delimiter=\"%s\">\n",
+    attr(x, "format"),
+    attr(x, "fence_type")
+  ))
+
+  cat_h2("Data")
+  print(x$data)
+  cat_h2("Body")
+  cat(x$body, "\n")
+  invisible(x)
+}
+
+cat_h2 <- function(x) {
+  cat(
+    strrep("\u2500", 4),
+    " ",
+    x,
+    strrep("\u2500", 4),
+    "\n",
+    sep = ""
   )
 }